requirements.txt
PyMySQL~=1.1.0
SQLAlchemy~=2.0.23
selenium~=4.15.1
ttkthemes~=3.2.2
python-dateutil~=2.8.2
requests~=2.31.0
validators~=0.22.0